+++ resources/js/router.js
@@ -5,7 +5,6 @@ import Race from '@/components/Pages/Race.vue'
 import Driver from '@/components/Pages/Driver.vue'
 import RaceList from '@/components/Pages/RaceList.vue'
 import DriverList from '@/components/Pages/DriverList.vue'
-import Statistics from '@/components/Pages/Statistics.vue'
 import Track from '@/components/Pages/Track.vue'
 import HotlapList from '@/components/Pages/HotlapList.vue'
 import TrackMedals from './components/Pages/TrackMedals.vue'
@@ -20,7 +19,6 @@ const routes = [
     { path: '/track/:id/hotlaps', component: HotlapList, name: 'TrackHotlaps' },
     { path: '/hotlap', component: HotlapList, name: 'HotlapList' },
     { path: '/trackMedals', component: TrackMedals, name: 'TrackMedals' },
-    { path: '/statistics', component: Statistics, name: 'Statistics' },
 ]
 
 const router = createRouter({